Use one ManiaPlanet folder for all ManiaPlanet games
In the compdata folder the games are linked via symlinks to the ManiaPlanet folder. This makes it possible to launch other trackmania games while running a mania planet game.
The renaming and resymlinking seems to work fast enough that the game doesn't crash. Tested on only 2 systems. This might need some extra testing.